+++ creme/feature_selection/__init__.py
@@ -1,4 +1,10 @@
+from .k_best import SelectKBest
 from .random import RandomDiscarder
+from .variance import VarianceThreshold
 
 
-__all__ = ['RandomDiscarder']
+__all__ = [
+    'RandomDiscarder',
+    'SelectKBest',
+    'VarianceThreshold'
+]