+++ tests/test_common.py
@@ -9,9 +9,11 @@ from sklearn.utils import estimator_checks
 from creme import cluster
 from creme import compat
 from creme import ensemble
+from creme import feature_selection
 from creme import linear_model
 from creme import naive_bayes
 from creme import preprocessing
+from creme import stats
 from creme import utils
 
 
@@ -29,7 +31,9 @@ ESTIMATORS = [
     preprocessing.MinMaxScaler(),
     preprocessing.MinMaxScaler() + preprocessing.StandardScaler(),
     preprocessing.StandardScaler() | linear_model.LinearRegression(),
-    preprocessing.PolynomialExtender()
+    preprocessing.PolynomialExtender(),
+    feature_selection.VarianceThreshold(),
+    feature_selection.SelectKBest(similarity=stats.PearsonCorrelation())
 ]